Refactor rename currentItemList to recallList
